Define Datum as an unsigned 64 bit value (#11872)
PG commit f90d23d0c518 introduced abbreviated keys sorting to macaddr
type. It transforms macaddr type to a raw Datum to compare. It generates
different results on PG & GP since the Datum is defined differently.

Datum was defined as a uintptr_t in PG, but int64 on GP. The commit
history doesn't show the reason why this has diverged from the upstream
at the beginning. But it seems to be a good thing to align this with
upstream.

Since CDB has an assumption of 8 bytes long Datum, and it requires to run
on x86_64 platform, using uintptr_t won't violate the original assumption on
the supported platforms.

Also, some regress tests are added to check the macaddr/uuid sorting
results.

+++ src/include/nodes/nodes.h
@@ -724,7 +724,7 @@ extern void outNode(struct StringInfoData *str, const void *obj);
 extern void outToken(struct StringInfoData *str, const char *s);
 extern void outBitmapset(struct StringInfoData *str,
 						 const struct Bitmapset *bms);
-extern void outDatum(struct StringInfoData *str, int64 value,
+extern void outDatum(struct StringInfoData *str, uintptr_t value,
 					 int typlen, bool typbyval);
 extern char *nodeToString(const void *obj);
 extern char *bmsToString(const struct Bitmapset *bms);
@@ -747,7 +747,7 @@ extern void *stringToNode(const char *str);
 extern void *stringToNodeWithLocations(const char *str);
 #endif
 extern struct Bitmapset *readBitmapset(void);
-extern int64 readDatum(bool typbyval);
+extern uintptr_t readDatum(bool typbyval);
 extern bool *readBoolCols(int numCols);
 extern int *readIntCols(int numCols);
 extern Oid *readOidCols(int numCols);
